requirements.txt
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
wandb>=0.15.0
ptflops>=0.7.0
torchmetrics>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0
